Do not save duplicated options
