+++ lib/forms/Interfaces.ts
@@ -1,0 +1,10 @@
+export type NewSelectOptionType = {
+  id?: string;
+  index?: number;
+  name: string;
+  color: string;
+};
+
+export type ExistingSelectOption = NewSelectOptionType & {
+  id: string;
+}
